# Conflicts:
#	RR-Inuring/src/main/java/com/scor/rr/service/InuringGroupingService.java
#	RR-Inuring/src/test/java/InuringInputNodeTest.java

+++ RR-Inuring/src/main/java/com/scor/rr/repository/SelectedPltRepository.java
@@ -4,7 +4,7 @@ package com.scor.rr.repository;
 import com.scor.rr.views.SelectedPLTView;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SelectedPltRepository extends JpaRepository<SelectedPLTView, Integer> {
+public interface SelectedPltRepository extends JpaRepository<SelectedPLTView, Long> {
 
-    SelectedPLTView findByPltId(int pltId);
+    SelectedPLTView findByPltId(long pltId);
 }
